chore(client): move network into internal/network
进一步将目录结构与MiraiGo对齐
