+++ client/packet.go
@@ -3,7 +3,7 @@ package client
 // from https://github.com/Mrs4s/MiraiGo/blob/master/client/packet.go
 
 import (
-	"github.com/LagrangeDev/LagrangeGo/client/network"
+	"github.com/LagrangeDev/LagrangeGo/client/internal/network"
 )
 
 func (c *QQClient) uniPacket(command string, body []byte) (uint32, []byte) {
